fix(glance): correct hacker news widget type and limit
The `type` for the hacker news widget was incorrect, causing
it to not render properly. This commit corrects the type
from `hackernews` to `hacker-news`. Also the limit on the
number of articles displayed has been increased to 10. The
collapse-after setting was added to collapse the entries
after the 5th one.